refactored
refactored with preprocessor macros to make code maintenance easier
