+++ inc/Kdict.h
