+++ inc/UContainer.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <utility>
 #include <stdio.h>
 #include <iostream>
 #include "globals.h"
@@ -11,12 +12,20 @@ namespace py = pybind11;
 
 typedef struct {
     uint8_t* suffixes;
+#if KDICT
     py::object* objs;
+#endif
     uint16_t size;
 } __attribute__ ((__packed__)) UC;
 
 void print( UC* uc, int k, int depth );
+
+#if KDICT
 void uc_insert( UC* uc, uint8_t* bseq, int k, int depth, int idx, py::object* obj );
+#elif KSET
+void uc_insert( UC* uc, uint8_t* bseq, int k, int depth, int idx );
+#endif
+
 void free_uc( UC* uc );
 int uc_contains( UC* uc, int k, int depth, uint8_t* bseq );
 int uc_find( UC* uc, int k, int depth, uint8_t* bseq );
