+++ inc/Vertex.h
@@ -1,14 +1,12 @@
 #pragma once
 
-#include "set/CContainer.h"
-#include "set/UContainer.h"
+#include <pybind11/pybind11.h>
+#include "CContainer.h"
+#include "UContainer.h"
 #include "globals.h"
 #include <jemalloc/jemalloc.h>
 
-/*typedef struct {
-    CC* cc;
-    UC uc;
-} __attribute__ ((__packed__)) Vertex;*/
+namespace py = pybind11;
 
 struct Vertex{
     CC* cc;
@@ -28,6 +26,13 @@ void init_vertex( Vertex* v );
 void free_vertex( Vertex* v );
 bool vertex_contains( Vertex* v, uint8_t* bseq, int k, int depth );
 void burst_uc( Vertex* v, int k, int depth );
+
+#if KDICT
+void vertex_insert( Vertex* v, uint8_t* bseq, int k, int depth, py::object* obj );
+#elif KSET
 void vertex_insert( Vertex* v, uint8_t* bseq, int k, int depth );
+#endif
+
+py::object* vertex_get( Vertex* v, uint8_t* bseq, int k, int depth );
