+++ src/TestLosePoint.cc
@@ -4,7 +4,7 @@
 #include <string>
 #include <assert.h>
 
-#include "set/Kcontainer.h"
+#include "Kcontainer.h"
 
 
 int main( int argc, char* argv[] )
