+++ src/UContainer.cc
@@ -1,8 +1,11 @@
-#include "set/UContainer.h"
+#include "UContainer.h"
 
 void init_uc( UC* uc )
 {
     uc->suffixes = NULL;
+#if KDICT
+    uc->objs = NULL;
+#endif
     uc->size = 0;
 }
 
@@ -19,12 +22,19 @@ void print( UC* uc, int k, int depth )
     }
 }
 
+#if KDICT
+void uc_insert( UC* uc, uint8_t* bseq, int k, int depth, int idx, py::object* obj )
+#elif KSET
 void uc_insert( UC* uc, uint8_t* bseq, int k, int depth, int idx )
+#endif
 {
     int len = calc_bk( k );
     if( uc->suffixes == NULL )
     {
         uc->suffixes = ( uint8_t* ) calloc( len , sizeof( uint8_t ) );
+#if KDICT
+        uc->objs = ( py::object* ) calloc( len , sizeof( py::object ) );
+#endif
     }
     else
     {
@@ -32,24 +42,40 @@ void uc_insert( UC* uc, uint8_t* bseq, int k, int depth, int idx )
                 uc->suffixes,
                 len * ( uc->size + 1 ) * sizeof( uint8_t )
                 );
+#if KDICT
+        uc->objs = ( py::object* ) realloc(
+                uc->objs,
+                ( uc->size + 1 ) * sizeof( py::object )
+                );
+#endif
     }
 
     if( uc->size < CAPACITY )
     {
-        //int idx = binary_search( uc->suffixes, uc->size, len, bseq );
-
         int bytes_to_move = ( uc->size - idx ) * len;
-        idx = idx * len;
+        int suffix_idx = idx * len;
         if( bytes_to_move > 0 )
         {
             std::memmove(
-                    &uc->suffixes[ idx + len ],
-                    &uc->suffixes[ idx ],
+                    &uc->suffixes[ suffix_idx + len ],
+                    &uc->suffixes[ suffix_idx ],
                     bytes_to_move
                     );
+            
+#if KDICT
+            bytes_to_move = ( uc->size - idx ) * sizeof( py::object );
+            std::memmove(
+                    &uc->objs[ idx + 1 ],
+                    &uc->objs[ idx ],
+                    bytes_to_move
+                    );
+#endif
         }
 
-        std::memcpy( &uc->suffixes[ idx ], bseq, len );
+#if KDICT
+        std::memcpy( &uc->objs[ idx ], obj, sizeof( py::object ) );
+#endif
+        std::memcpy( &uc->suffixes[ suffix_idx ], bseq, len );
         uc->size++;
     }
 }
@@ -59,6 +85,9 @@ void free_uc( UC* uc )
     if( uc->suffixes != NULL )
     {
         free( uc->suffixes );
+#if KDICT
+        free( uc->objs );
+#endif
     }
 }
 
@@ -67,10 +96,19 @@ void uc_remove( UC* uc, int bk, int idx )
     int suffix_idx = idx * bk;
     int bytes_to_move = ( uc->size - ( idx + 1 ) ) * bk;
     std::memmove(
-            &uc->suffixes[ idx ],
-            &uc->suffixes[ idx + bk ],
+            &uc->suffixes[ suffix_idx ],
+            &uc->suffixes[ suffix_idx + bk ],
             bytes_to_move
             );
+#if KDICT
+    bytes_to_move = ( uc->size - ( idx + 1 ) ) * sizeof( py::object );
+    std::memmove(
+            &uc->objs[ idx ],
+            &uc->objs[ idx + 1 ],
+            bytes_to_move
+            );
+#endif
+
     uc->size--;
 }
 
@@ -81,34 +119,14 @@ int uc_find( UC* uc, int k, int depth, uint8_t* bseq )
         return uc->size;
     }
 
-    int idx = binary_search_contains( uc->suffixes, uc->size, calc_bk( k ), bseq );
-    if( idx > -1 )
+    std::pair< bool, int > res = binary_search( uc->suffixes, uc->size, calc_bk( k ), bseq );
+    if( res.first )
+    {
+        return res.second;
+    }
+    else
     {
         return uc->size;
-    }
-    else
-    {
-        idx = ( idx + 1 ) * -1;
-        return idx;
-    }
-
-}
-
-int uc_contains( UC* uc, int k, int depth, uint8_t* bseq )
-{
-    if( uc->suffixes == NULL )
-    {
-        return 0;
-    }
-
-    int idx = binary_search_contains( uc->suffixes, uc->size, calc_bk( k ), bseq );
-    if( idx > -1 )
-    {
-        return idx;
-    }
-    else
-    {
-        return -1;
     }
 }
